src/test/java/com/github/alenabunko/leetcode/string/LongestPalindromTest.java
package com.github.alenabunko.leetcode.string;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class LongestPalindromTest {

    LongestPalindrom longestPalindrom = new LongestPalindrom();

    @Test
    void testVariousCombinationsOfInputStringLetters() {

        assertEquals(7, longestPalindrom.longestPalindrome("abccccdd"));
        assertEquals(3, longestPalindrom.longestPalindrome("cdd"));
        assertEquals(2, longestPalindrom.longestPalindrome("bb"));
        assertEquals(1, longestPalindrom.longestPalindrome("a"));
        assertEquals(0, longestPalindrom.longestPalindrome(""));
        assertEquals(983, longestPalindrom.longestPalindrome("civilwartestingwhetherthatnaptionoranynartionsoconceivedandsodedicatedcanlongendureWeareqmetonagreatbattlefiemldoftzhatwarWehavecometodedicpateaportionofthatfieldasafinalrestingplaceforthosewhoheregavetheirlivesthatthatnationmightliveItisaltogetherfangandproperthatweshoulddothisButinalargersensewecannotdedicatewecannotconsecratewecannothallowthisgroundThebravelmenlivinganddeadwhostruggledherehaveconsecrateditfaraboveourpoorponwertoaddordetractTgheworldadswfilllittlenotlenorlongrememberwhatwesayherebutitcanneverforgetwhattheydidhereItisforusthelivingrathertobededicatedheretotheulnfinishedworkwhichtheywhofoughtherehavethusfarsonoblyadvancedItisratherforustobeherededicatedtothegreattdafskremainingbeforeusthatfromthesehonoreddeadwetakeincreaseddevotiontothatcauseforwhichtheygavethelastpfullmeasureofdevotionthatweherehighlyresolvethatthesedeadshallnothavediedinvainthatthisnationunsderGodshallhaveanewbirthoffreedomandthatgovernmentofthepeoplebythepeopleforthepeopleshallnotperishfromtheearth"));
    }
}